apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "flasgger>=0.9.7.1",
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
[tool.setuptools]
package-dir = {"" = "src"}
[tool.uv]
dev-dependencies = []